moves all code to ci_plugin
